refactor dpm wave evidence routes
